refactor(esp32): rename ble_nus_client to ble_nus for cleaner naming
- Rename ble_nus_client.c/.h to ble_nus.c/.h
- Update all includes from "ble_nus_client.h" to "ble_nus.h"
- Update CMakeLists.txt to reference new filenames
- No functional changes, purely naming cleanup

Since the unused NUS GATT server code was removed, the "client" suffix
is no longer needed - the remaining NUS code is clearly for client functionality.
This provides cleaner, more intuitive naming:

Before: ble_nus_client.c (NUS client) + ble_nus.c (unused server)
After:  ble_nus.c (NUS client only)
